% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_current_programme}
\alias{get_current_programme}
\title{Get programme for today}
\usage{
get_current_programme()
}
\value{
List of data.frames with programme (a data.frame for each station)
}
\description{
Get programme for today
}
\examples{
\dontrun{
get_current_programme()
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
