% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_current_station_programme}
\alias{get_current_station_programme}
\title{Get programme for today for a given station}
\usage{
get_current_station_programme(station)
}
\arguments{
\item{station}{ID of station}
}
\value{
A data.frames with programme
}
\description{
Get programme for today for a given station
}
\examples{
\dontrun{
get_current_station_programme("radiozurnal")
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
