% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_day_programme}
\alias{get_day_programme}
\title{Get programme for a given day for all stations}
\usage{
get_day_programme(day)
}
\arguments{
\item{day}{Date (in YYYY-MM-DD format)}
}
\value{
List of data.frames with programme (one data.frame for each station)
}
\description{
Get programme for a given day for all stations
}
\examples{
\dontrun{
get_day_programme("2021-09-30")
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
