% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_day_station_programme}
\alias{get_day_station_programme}
\title{Get programme for a given day and station}
\usage{
get_day_station_programme(day, station)
}
\arguments{
\item{day}{Date (in YYYY-MM-DD format)}

\item{station}{ID of station}
}
\value{
A data.frame with programme
}
\description{
Get programme for a given day and station
}
\examples{
\dontrun{
get_day_programme("2021-09-30", "radiozurnal")
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
