% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_day_wave_playlist}
\alias{get_day_wave_playlist}
\title{Get Radio Wave's playlist for a given day}
\usage{
get_day_wave_playlist(day)
}
\arguments{
\item{day}{Date (in YYYY-MM-DD format)}
}
\value{
A data.frame with playlist
}
\description{
Get Radio Wave's playlist for a given day
}
\examples{
\dontrun{
get_day_wave_playlist("2021-09-30")
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
