% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_stations}
\alias{get_stations}
\title{Get stations}
\usage{
get_stations()
}
\value{
Data.frame of data on stations (id, name, description, links)
}
\description{
Get stations
}
\examples{
\dontrun{
get_stations()
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
