% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_wave_playlist}
\alias{get_wave_playlist}
\title{Get Radio Wave's playlist for today}
\usage{
get_wave_playlist()
}
\value{
A data.frame with playlist
}
\description{
Get Radio Wave's playlist for today
}
\examples{
\dontrun{
get_wave_playlist()
}
}
\seealso{
\url{https://data.irozhlas.cz/opendata/}
}
