% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_twodigit_date}
\alias{make_twodigit_date}
\title{Make single digit dates (days, months) to two digits}
\usage{
make_twodigit_date(x)
}
\arguments{
\item{x}{number}
}
\description{
Make single digit dates (days, months) to two digits
}
